Add components with steps to avoid to much code in one file
