+++ view/frontend/web/js/components/steps/BaseSuccessStep.vue
@@ -1,0 +1,22 @@
+<template>
+    <section v-if="step === 'success'">
+        We did it!
+
+        <a href="/what-is-new.html">
+            Back to category
+        </a>
+    </section>
+</template>
+
+<script>
+export default {
+    computed: {
+        cartId() {
+            return this.$store.state.cart;
+        },
+        step() {
+            return this.$store.state.step;
+        }
+    },
+}
+</script>
